v1
known "issue": it's possible for their to be separate "parts" to the structure.  For example, there could be a (bishop & pawn) structure on the right side of board and (rook & rook) structure on the left side, not protecting each other.    At this crossroads in the project, my initial idea was to create something to check if there was a chain of protection, like a pawn protecting a rook, that protects a bishop, ... and make sure the chain is long enough, but had a better idea (felt this solution would be a bit too complex/clunky anyway).  In the next version, I'm going to count the number of "protections" in each generated structure, and over time as it goes through more and more, it will save the board with the largest value.  By doing this, I'm essentially optimizing the board for the number of protections it has, and thus getting a more connected board, as a board will need to be very well connected to have that many protections.