final code update
This update implements the feature that I mentioned last time, where it maximizes the number of "protections" of a self protected structure.

An example: if you have a rook, a bishop and a pawn on a board, where the rook protects the pawn and the bishop protects both the rook and the pawn, the pawn is being protected twice.  So in this update, I calculate a value for each structure that is deemed "self protective" (all pieces are protected and protecting at least one other piece), where the value represents the number of "protections".  As a result, this will make a board that is extremely well protected and tied together, thus creating the final "Frishberg Chess Structure".

Note: the output in this update has 29 protections